Add bin_cmd argument to some modules
The modules which require access to the rabbitmq-plugins
and rabbitmqctl binary may need to use a custom path, or
perhaps use a container.

This patch adds the capability in these modules to provide
a binary path or any command which will yield the required
binary. This enables the use of a container to access it.