Change variable name
The variable `change_required` initially confused me; it doesn't really
say whether or not changes are required to the queue, but instead
describes whether the queue's presence matches its desired presence
(i.e. is it there if we want it there?). `change_required` would
probably better describe the operation we're doing in the next lines,
of figuring out if any of the properties differ from what we specify.