Refactor checks into function; add verbose output
Previously, the module wrote out all of the checks to see if any of the
properties of the queue had changed, evaluated them all at once, and
only printed a generic failure method.

Also, previously, specifying an option in the arguments dict would
result in it being created, but the checks didn't catch it, so despite
the options not changing, the check for changed attributes would cause
the module to fail the second time it was run.

This commit fixes both of those issues. Rather than synchronizing the
module.params dict and the module.params['arguments'] dict after the
checks are done, it does before, and it modifies the checks to be based
off module.params['arguments'], which eventually becomes the final
source of truth. This lets the checks run against
module.params['arguments'] rather than module.params.

This commit also extracts the check logic to a function, which describes
what failed and why.